Match key floor to built unit; fp16 commit domain; item-7 Vcorr goldens
Issue #7, single superset regen (job 53047303, verifier fatal=0):
- _quantize_risk_key now implements the hardware map literally: RNE-12
  at each value's own exponent, POST-round membership, ebase =
  e_pre(max) - 62, disjoint bottom bin, natural anchor carry. Zero mask
  movement on the 12 rows vs the previous floor.
- Commit test pinned to the 2-byte sidecar domain (strict fp16 < fp16)
  in the key-domain golden path, the run() selection block, and the GPU
  runner; item-1..5 fields verified byte-identical under the switch.
- New item-7 fields: full-context commit sidecars, per-probe Vcorr band
  partials in dual operand domains (_ref fp64 operands / _hw one-fp16-
  cast-per-token-dim-tier off the dualplane reconstruction), per-token
  exact-set operands with both planes, K-move fixup vectors on all 3 kd
  rows. Writer self-checks: per-domain band-sum decomposition,
  eps_band-bounded trace consistency (measured max 4.2e-7), dualplane
  reconstruction bound (measured 3.1e-5). Measured operand-domain dv
  divergence: max 4.0e-7, no record within 25x of the 1e-5 flag line.
- Verifier split: moved item-6 outputs report-only, everything else
  byte-asserted; dv_hw-dv_ref and dv_trace_delta measured and flagged.

+++ benchmark/selector_eval/runners/hf_paged_pq_intervention_joint_one_group.py
@@ -39,7 +39,7 @@ from benchmark.selector_eval.runners.hf_paged_pq_intervention_joint_vprefix impo
 # OPEN-2/M6): the top 10% of the K ranked prefix / V risk-ranked exact set
 # reads full precision; the rest reads the per-row absmax-int8 plane, with
 # V lo reads additionally gated by the per-token commit test
-# (int8_err < code_error). Mirrors --precision_k_hi_frac 0.1
+# (fp16(int8_err) < fp16(code_error)). Mirrors --precision_k_hi_frac 0.1
 # --precision_v_hi_frac 0.1 --precision_lo_mode int8 --precision_lo_bits 8
 # in run_joint_kv_budget_policy_eval.py — keep in sync.
 _FROZEN_PRECISION_K_HI_FRAC = 0.1
@@ -972,14 +972,15 @@ def process_one_joint_kv_head(runtime, kv_head_i: int) -> bool:
         scores_lo_h_t = (queries_h @ _rowwise_int8_qdq(keys32_t).transpose(0, 1)) / sqrt_dim
         del keys32_t
         # V lo tier: int8-plane values, per-token commit test against the
-        # V-PQ code-error stat (both fp64, matching the CPU reference
-        # compare), and the committed lo residual pre-zeroed on failed
-        # commits so the V grid can fold it into one cumsum.
+        # V-PQ code-error stat in the 2-byte sidecar domain (fp16 int8-error
+        # vs fp16 code-error, matching the CPU reference compare), and the
+        # committed lo residual pre-zeroed on failed commits so the V grid
+        # can fold it into one cumsum.
         values32_t = values_t.to(device=device, dtype=torch.float32)
         values_lo_t = _rowwise_int8_qdq(values32_t)
         int8_err_t = (values32_t - values_lo_t).pow(2).sum(dim=1, dtype=torch.float64)
         del values32_t
-        v_commit_mask_t = int8_err_t < code_error_t.to(dtype=torch.float64)
+        v_commit_mask_t = int8_err_t.to(dtype=torch.float16) < code_error_t.to(dtype=torch.float16)
         residual_lo_commit_t = torch.where(
             v_commit_mask_t.reshape(-1, 1),
             values_lo_t - vhat_all_t.to(dtype=torch.float32),
